# specs/001-ticket-dedup-dashboard/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DedupTickets API
  description: API for the Ticket Deduplication & Clustering Dashboard
  version: 1.0.0
  contact:
    name: DedupTickets Team
servers:
  - url: /api/v1
    description: API v1
security:
  - ApiKeyAuth: []
...
paths:
...
  /trends/top-drivers:
    get:
      tags: [Trends]
      summary: Get top recurring drivers
      operationId: getTopDrivers
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: week
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Top drivers ranked by frequency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopDriversResponse'
  /trends/fastest-growing:
    get:
      tags: [Trends]
      summary: Get fastest growing drivers
      operationId: getFastestGrowing
      parameters:
        - name: compare_period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: week
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Drivers sorted by growth rate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrowingDriversResponse'
  /trends/most-duplicated:
    get:
      tags: [Trends]
      summary: Get drivers with highest duplication rate
      operationId: getMostDuplicated
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: week
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Drivers ranked by duplication rate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuplicatedDriversResponse'

  # ==================== AUDIT ====================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    Driver:
      type: object
      required: [id, theme_summary, cluster_count, ticket_count]
      properties:
        id:
          type: string
          format: uuid
        theme_summary:
          type: string
        matching_pattern:
          type: object
        cluster_count:
          type: integer
        ticket_count:
          type: integer
        duplication_rate:
          type: number
        trend_direction:
          type: string
          enum: [up, down, stable]
        trend_percentage:
          type: number
        last_seen_at:
          type: string
          format: date-time
    TopDriversResponse:
      type: object
      required: [data, period]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
        period:
          type: string
    GrowingDriversResponse:
      type: object
      required: [data, compare_period]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
        compare_period:
          type: string
    DuplicatedDriversResponse:
      type: object
      required: [data, period]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Driver'
        period:
          type: string

    # ==================== AUDIT ====================
...